Operations Risk and Financial Risk/RBI_RiskGuidanceGPT/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-text-splitters
langchain-community
faiss-cpu
pypdf
pdfplumber
python-dotenv
sentence-transformers
